Notes for information regarding to Cyber attacks
To whom it may be concern regards to more about Cyber attacks or "Hackers" better know as commonly people refer to. This will hopefully provide information on my
tricks and tips i hade to pick up and learn myself along the way, youtube will has a very small selection to pick from regards to certain information on really mastering the fundamentals of true masters of hacking. Helping you understand if any on how computer work and how theses functions can be manipulated to your will as you see fit to do so for what ever the situation may be.